User Authentication - Users can signup
Added a user model and updated database with user table. Created routes to go to signup page and to post to users (add new user to database). Created a users controller with new and create methods. Created a view for the signup page.

When a user signs up, their info is put in the database with hashed password. If the user signup fails the user is redirected to the signup page.