Add Order Details
The order page now populates with the product information for the order ID.

Connected the product model to the line_item model with a has many relationship. Created a new method in the order controller that looks up all the products in a given order as an array by connecting through the line_items table. This function also passes the order quantity info from line_items for each product. This product array is passed to the order view as an instance variable where it is used.

The username and email is also on the page. Currently those are associated with the current user, as the order DB is not taking correct email information. This will be fixed soon.