Fix Empty Cart Bug
The cart page now renders a notification that the cart is empty when a user has not items in their cart. The order table and purchase button will not render in this case.

Also added a guest name to the page when a user is not logged in, so that Guest appears in the strip purchase window.