Feature Spec - Add to Cart
Tests that when a user clicks add to cart from the main page, that an item is added to cart. Checks by inspecting that the 'My Cart(count)' increments in the nav.

Also added some helper css classes for easier more robust lookups in erb files. And refactored previous tests accordingly.