Feature Spec - Users can login
Verify that a user with correct credentials can login and will redirect to main page. A test user is created, their credentials are filled into the forms, and the Login button is clicked. Redirect is verified by checking for existence of username in the nav and the existence of the section.products-index css selector.